Move comm functions to communication.ts
